Added app functions for modularity to k8helmpipeline
